+++ generators/biski16.c
@@ -33,9 +33,9 @@ static inline uint64_t get_bits_raw(void *state)
 static void *create(const CallerAPI *intf)
 {
     Biski16State *obj = intf->malloc(sizeof(Biski16State));
-    obj->loop_mix = intf->get_seed64();
-    obj->mix = intf->get_seed64();
-    obj->ctr = intf->get_seed64();
+    obj->loop_mix = (uint16_t) intf->get_seed64();
+    obj->mix = (uint16_t) intf->get_seed64();
+    obj->ctr = (uint16_t) intf->get_seed64();
     return obj;
 }
