+++ generators/melg44497.c
@@ -61,14 +61,15 @@ void MelgState_init(MelgState *obj, uint64_t seed)
     uint64_t *x = obj->melg;
     obj->melg[0] = seed;
     for (obj->pos = 1; obj->pos < NN; obj->pos++) {
-        x[obj->pos] = (6364136223846793005ULL * (x[obj->pos-1] ^ (x[obj->pos-1] >> 62)) + obj->pos);
+        x[obj->pos] = (6364136223846793005ULL * (x[obj->pos-1] ^ (x[obj->pos-1] >> 62)) + (unsigned int) obj->pos);
     }
-    obj->lung = (6364136223846793005ULL * (x[obj->pos-1] ^ (x[obj->pos-1] >> 62)) + obj->pos);
+    obj->lung = (6364136223846793005ULL * (x[obj->pos-1] ^ (x[obj->pos-1] >> 62)) + (unsigned int) obj->pos);
     obj->pos = 0;
     obj->function_p = case_1;
 }
 
-static inline uint64_t case_generic(MelgState *obj, int sh1, int sh2, int sh3,
+static inline uint64_t case_generic(MelgState *obj,
+    int sh1, int sh2, int sh3,
     uint64_t (*new_function_p)(MelgState *obj))
 {
     uint64_t *st = obj->melg;
